# workload/internal/httpapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Report Workshop",
    "version": "1.0.0",
    "description": "Synthetic document analysis. Local demo endpoints require private access; cloud authentication remains a platform concern."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:18080",
      "description": "Local Compose"
    }
  ],
  "paths": {
    "/healthz": {
      "get": {
        "operationId": "health",
        "responses": {
          "200": {
            "description": "Process responds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            }
          }
        }
      }
    },
    "/readyz": {
      "get": {
        "operationId": "ready",
        "responses": {
          "200": {
            "description": "Database responds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            }
          },
          "503": {
            "description": "Database unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs": {
      "post": {
        "operationId": "submitJob",
        "description": "Request body is bounded to 4096 bytes. An identical idempotency key/request returns the original job; a different request conflicts.",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,128}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Submission"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted or existing job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "Job status path",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid key or request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Key reused with different request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Storage unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{id}": {
      "get": {
        "operationId": "getJob",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Job identifier returned by submission."
          }
        ],
        "responses": {
          "200": {
            "description": "Durable job state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "404": {
            "description": "Unknown job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Storage unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{id}/report": {
      "get": {
        "operationId": "getReport",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Job identifier returned by submission."
          }
        ],
        "responses": {
          "200": {
            "description": "Completed deterministic report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Report"
                }
              }
            }
          },
          "404": {
            "description": "Unknown job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Report not committed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Storage unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "operationId": "getSpecification",
        "responses": {
          "200": {
            "description": "This OpenAPI specification",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Submission": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "fixture",
          "algorithm"
        ],
        "properties": {
          "fixture": {
            "type": "string",
            "enum": [
              "tiny-v1",
              "batch-v1"
            ]
          },
          "algorithm": {
            "const": "tokens-v1"
          }
        }
      },
      "Job": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "fixture",
          "algorithm",
          "state",
          "attempt",
          "created_at",
          "completed_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[0-9a-f]{32}$"
          },
          "fixture": {
            "type": "string",
            "enum": [
              "tiny-v1",
              "batch-v1"
            ]
          },
          "algorithm": {
            "const": "tokens-v1"
          },
          "state": {
            "enum": [
              "pending",
              "running",
              "succeeded",
              "failed"
            ]
          },
          "attempt": {
            "type": "integer",
            "minimum": 0,
            "maximum": 5
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "completed_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
      "Report": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "documents",
          "tokens",
          "unique_tokens",
          "duplicate_documents",
          "input_sha256"
        ],
        "properties": {
          "documents": {
            "type": "integer",
            "minimum": 0
          },
          "tokens": {
            "type": "integer",
            "minimum": 0
          },
          "unique_tokens": {
            "type": "integer",
            "minimum": 0
          },
          "duplicate_documents": {
            "type": "integer",
            "minimum": 0
          },
          "input_sha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$"
          }
        }
      },
      "Health": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "enum": [
              "ok",
              "ready"
            ]
          }
        }
      },
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "enum": [
              "invalid_idempotency_key",
              "invalid_request",
              "idempotency_conflict",
              "storage_unavailable",
              "database_unavailable",
              "job_not_found",
              "report_not_ready"
            ]
          }
        }
      }
    }
  }
}